+++ polkadot/runtime/parachains/src/runtime_api_impl/mod.rs
@@ -26,5 +26,5 @@
 //! 2. Move methods from `vstaging` to `v3`. The new stable version should include all methods from
 //!    `vstaging` tagged with the new version number (e.g. all `v3` methods).
 
-pub mod v10;
+pub mod v11;
 pub mod vstaging;
